Working read, buggy write
With the enourmous help of Voginator, the program can read key events
and output them with NSpireIO. All credit to Voginator. He found the key
press hooks for my version of the TI-Nspire OS (4.2.0.523).

But NspireIO outputs to the serial GPIO pin rather than the USB port
like Lua, so either I have to port the Lua ASI library or read up on the
NavNet API.